Added custom exceptions and improved sevenz.

+++ klgists/files/delete_hidden_files.py
@@ -1,9 +1,11 @@
 import os
+from typing import Set, Optional
 from klgists.files.scantree import scan_for_files
 
-def delete_hidden_files(directory: str) -> None:
+def delete_hidden_files(directory: str, filename_starts: Optional[Set[str]] = None) -> None:
 	"""Deletes any files beginning with '.' or '~'"""
+	if filename_starts is None: filename_starts = {'.', '~'}
 	for f in scan_for_files(directory):
-		if os.path.basename(f).startswith('.') or os.path.basename(f).startswith('~'):
-			os.remove(f)
+		for s in filename_starts:
+			if os.path.basename(f).startswith(s): os.remove(f)
